Add support for NotMapped attribute
Currently all public properties of a class are mapped to columns in the
database table.  EF uses the [NotMapped] attribute to signal that a
property should not be mapped into the database.  Adding a check for
this attribute allows the data class to control in a standard way what
properties belong to the database.